Always use the quest history obtained from GetQuestsCompleted()
The variable CodexHistory has been removed. Because GetQuestsCompleted() is always available, the addon does not need to save the player's quest history by itself. At the same time, there is no longer a need to monitor the player for abandoning quests.

In addition, this will fix the available quests display for players who are starting to use the addon midway.

A new variable CodexHiddenQuests has been added to save the quests hidden by the player.